# Improve decoding performance.